Modularize API so that each story lives in its own path (and its source code lives in its own folder).
