+++ package.json
@@ -2,7 +2,7 @@
   "scripts": {
     "build": "tsc",
     "lint": "node_modules/.bin/eslint src --ext .ts",
-    "serve": "node dist/server.js",
+    "serve": "node dist/main.js",
     "start": "npm run serve"
   },
   "dependencies": {
@@ -24,6 +24,7 @@
     "express-session": "^1.17.2",
     "mysql2": "^2.3.3",
     "nanoid": "^3.3.2",
+    "path": "^0.12.7",
     "sequelize": "^6.16.1",
     "typescript": "^4.6.3",
     "uuid": "^8.3.2"
