+++ src/stories/hubbles_law/main.ts
@@ -1,0 +1,4 @@
+import router from "./router";
+
+export const ROUTE_PATH = "/hubbles_law";
+export const ROUTER = router;
