+++ src/stories/hubbles_law/models/galaxy.ts
